Assembly version with BMI1 instructions
Up to GCC 7.2, the compiler produces following code for
the main loop of bitmap_decode_ctz:

  400d80:       31 f6                   xor    %esi,%esi
  400d82:       48 83 c0 01             add    $0x1,%rax
  400d86:       f3 48 0f bc f1          tzcnt  %rcx,%rsi
  400d8b:       c4 e2 f0 f3 c9          blsr   %rcx,%rcx
  400d90:       44 01 c6                add    %r8d,%esi
  400d93:       48 85 c9                test   %rcx,%rcx
  400d96:       89 74 82 fc             mov    %esi,-0x4(%rdx,%rax,4)

I'm not an expert, but at least two instructions are not needed
here. The first xor %esi, %esi seems superfluous, as the %esi
is overwritten later by tzcnt (it might be used output dependency
breaker).

More important is not taking into account that blsr sets ZF when
result is zero. Thanks to that test %rcx, %rcx wouldn't be needed
if the add was moved before blsr.